+++ assets/js/pages/contact.js
@@ -1,1 +1,2 @@
 import "../../styles/pages/contact.scss";
+import "../utility/mobile-navbar";
