+++ assets/js/pages/project-view.js
@@ -1,1 +1,2 @@
 import "../../styles/pages/project/project-view.scss";
+import "../utility/mobile-navbar";
