+++ assets/js/pages/projects.js
@@ -1,5 +1,6 @@
 import "../../styles/pages/project/projects.scss";
 import TomSelect from 'tom-select';
+import "../utility/mobile-navbar";
 
 /* Selectize part */
 const select = document.querySelectorAll('.tom-select');
