+++ assets/styles/partials/_navbar.scss
@@ -94,6 +94,8 @@ nav {
         padding: 15px;
         border-bottom: 1px solid #f2f2f2;
         text-align: center;
+        display: flex;
+        justify-content: center;
 
         &:first-child {
           border-top: 1px solid #f2f2f2;
@@ -113,6 +115,29 @@ nav {
       padding: 0 15px;
       box-sizing: border-box;
       border-right: 1px solid #f2f2f2;
+
+      .language {
+        display: flex;
+        align-items: center;
+        margin-left: 15px;
+        text-wrap: nowrap;
+
+        @media (max-width: $tablet) {
+          margin-left: 0;
+        }
+
+        img {
+          margin-left: 5px;
+          height: 20px;
+          width: auto;
+          object-fit: cover;
+          border-radius: 5px;
+          @media (max-width: $tablet) {
+            margin-left: 0;
+            height: 30px;
+          }
+        }
+      }
 
       &:last-child {
         border-right: none;
