+++ config/packages/translation.yaml
@@ -3,6 +3,5 @@ framework:
     translator:
         default_path: '%kernel.project_dir%/translations'
         fallbacks:
-            - en
             - fr
-        providers:
+            - en
